Exit gracefully on signals
Neither fprintf nor exit are async-signal-safe.  Instead of exiting in
the signal handler, set a flag to be checked in the main loop, and clean
up after the loop.  If we receive a second signal before exiting, exit
immediately without cleaning up.

We no longer catch SIGABRT.  If we called abort the process is already
in an inconsistent state, so we should just dump core.